requirements.txt
openai>=1.4.0
python-dotenv>=1.0.0
langchain>=0.1.20
langchain-openai>=0.1.6